Add files via upload
The most favorable architecture design for the BCMessaging software is the client-server architecture pattern because it is able to provide and deliver several of the same functions to millions of users across a network using separate servers. Each of the servers is delegated to a certain task in the application, and corresponding to the fact that there are very few functions related to a messaging software, it would be simple to implement. Additionally, in this design, general functions, such as sending and receiving messages, can be made available to all clients involved and does not need to be implemented by individual servers. Instead, the users would be able to engage in instant messaging through the central server using each of their local machines, smartphones in this case, directly. Users can also add contacts through the central server. The client-server architecture pattern is the best suited to handling a constant stream of function calls from a large amount of users in different locations. Data from the BCMessaging shared database, namely the phone number database, should be accessible from anywhere. This database has few aspects of data manipulation associated with it, suggesting that the MVC, layered, repository, and pipe and filter patterns with their various data organization strategies are unneeded. When the communication between clients is the most important feature of the software and there are not too many services to manage, the client-server architecture is indubitably the best software architecture pattern to use.